fix(seo): serve lang="ja" on the Japanese pages
/japan and /academy/japan/* declare a self-referencing ja-JP hreflang while
the document reported lang="en", which Ahrefs flags as an hreflang/html-lang
mismatch (9 URLs) and which misreports the content language to screen readers
and translation tooling.

Next.js only lets the root layout render <html>, and its attributes are
serialized before any child renders, so one root layout cannot vary lang by
route. Split the root layout per document language instead:

- components/RootDocument.tsx holds the <html>/<body> shell, fonts, providers,
  and analytics, parameterized by lang.
- app/(en)/layout.tsx and app/(ja)/layout.tsx are the two root layouts. Route
  groups do not affect URLs, so every path is unchanged.
- /japan and /academy/japan/* move under app/(ja); everything else moves under
  app/(en). app/api stays put — route handlers use no layout.
- lib/site-metadata.ts holds the site-wide metadata both root layouts export.
- The 404 body moves to components/NotFoundContent.tsx: each root group gets a
  not-found.tsx wrapped by its own layout, and app/not-found.tsx renders its
  own document shell because Next serves it outside every layout.

scripts/sync-design-tokens.js parsed the next/font definitions out of
app/layout.tsx; repointed it at components/RootDocument.tsx so public/design.md
keeps resolving Inter, Geist Mono, and F37 Analog.

Fumadocs' UI locale stays "en" — no Japanese translation bundle is registered,
so only the document language varies.
