+++ app/(en)/(pricing)/pricing-self-host/page.tsx
