+++ app/(en)/(pricing)/pricing/page.tsx
