+++ app/(en)/[section]/[[...slug]]/page.tsx
