+++ app/(en)/[section]/layout.tsx
