+++ app/(en)/academy/(en)/[[...slug]]/page.tsx
