+++ app/(en)/academy/(en)/layout.tsx
