+++ app/(en)/blog/[...slug]/page.tsx
