+++ app/(en)/blog/layout.tsx
