+++ app/(en)/blog/page.tsx
