+++ app/(en)/faq/[[...slug]]/page.tsx
