+++ app/(en)/integrations/[[...slug]]/page.tsx
