+++ app/(en)/not-found.tsx
@@ -1,0 +1,5 @@
+import { NotFoundContent } from "@/components/NotFoundContent";
+
+export default function NotFound() {
+  return <NotFoundContent />;
+}
