+++ app/(en)/resources/[[...slug]]/page.tsx
