+++ app/(en)/resources/layout.tsx
