+++ app/(en)/self-hosting/[[...slug]]/page.tsx
