+++ app/(en)/workshop/[[...slug]]/page.tsx
