+++ app/(ja)/academy/japan/[[...slug]]/not-found.tsx
