+++ app/(ja)/academy/japan/[[...slug]]/page.tsx
