+++ app/(ja)/academy/japan/layout.tsx
