+++ app/(ja)/japan/page.tsx
