+++ app/not-found.tsx
@@ -1,29 +1,14 @@
-import Link from "next/link";
-import { NotFoundAnimation } from "@/components/NotFoundAnimation";
-import { BrokenLinkIssue } from "@/components/BrokenLinkIssue";
-import { HomeLayout } from "@/components/layout";
-import { Button } from "@/components/ui";
+import { RootDocument } from "@/components/RootDocument";
+import { NotFoundContent } from "@/components/NotFoundContent";
 
+/**
+ * 404 for URLs that match no route in either root layout group. Next.js renders
+ * this one outside every layout, so it has to supply the document shell itself.
+ */
 export default function NotFound() {
   return (
-    <HomeLayout showAside={false} forceLight={false}>
-      <div className="flex flex-col items-center justify-center text-center sm:py-20 min-h-[calc(100vh-4rem)]">
-        <NotFoundAnimation />
-        <div className="flex flex-col gap-6 justify-center items-center">
-          <div className="flex flex-col gap-2 justify-center items-center">
-            <h1 className="mt-6 text-2xl font-bold">404: Page Not Found</h1>
-            <p className="mt-2 text-muted-foreground">
-              The page you were looking for does not exist.
-            </p>
-          </div>
-          <div className="flex flex-row flex-wrap gap-3 justify-center items-center">
-            <Button size="default" href="/">
-              Go back home
-            </Button>
-            <BrokenLinkIssue />
-          </div>
-        </div>
-      </div>
-    </HomeLayout>
+    <RootDocument lang="en">
+      <NotFoundContent />
+    </RootDocument>
   );
 }
